backend/requirements.txt
# Core API
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Ciencia de datos
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
xgboost==2.1.1
joblib==1.4.2

# Utilidades
requests==2.32.3
httpx==0.27.0
python-dotenv==1.0.1
